refactored project into more than one py file for possibly better maintainability
