+++ .gitignore
@@ -1,0 +1,6 @@
+
+.idea/
+build/
+__pycache__/
+dist/
+HuntOverlay.spec
